Moved Questioenr Stuff to legacy & Removed all references to Questioner code
