+++ legacy/old-questioner-code/AllyQuestionTask.java
