+++ legacy/old-questioner-code/NationAllyTask.java
